Changed url from '/card/characters' to '/character/card'
