+++ api/controllers/info/characters.js
@@ -46,23 +46,15 @@ exports.characters_get_all_characters = (req, res, next) => {
             weaponType: query.weapon_type,
             faction: query.faction_name,
             request: {
+              card: {
+                type: "GET",
+                message: "Access general information of this card",
+                url: process.env.API_DEV + "/characters/card/" + query.id
+              },
               info: {
                 type: "GET",
                 message: "Access more detailed information of this card",
-                url: process.env.API_DEV + "/info/characters/" + query.id
-              },
-              faction: {
-                type: "GET",
-                message: "Access faction data",
-                url: process.env.API_DEV + "/info/factions/" + query.faction_id
-              },
-              weaponType: {
-                type: "GET",
-                message: "Access Weapon Type data",
-                url:
-                  process.env.API_DEV +
-                  "/info/weaponType" +
-                  query.weapon_type_id
+                url: process.env.API_DEV + "/characters/info/" + query.id
               }
             }
           };
@@ -134,128 +126,4 @@ exports.characters_get_character = (req, res, next) => {
       return res.status(404).json({ message: "Character not found" });
     }
   });
-};
-
-/**
- * name: String (req)
- * weapon_type_id: id (req)(ref: WeaponType)
- * img_url: String (req)
- * quote: String (req)
- * faction_id: id (req) (ref: Faction)
- * story: String (def: "")
- * stats: {
- *      swp: Number (req)
- *      shp: Number (req)
- *      effectName: String (def: "")
- *      effectDesc: String (def: "")
- * }
- */
-
-exports.characters_create_character = (req, res, next) => {
-  // Security Measure
-  const {
-    name,
-    weapon_type_id,
-    img_url,
-    quote,
-    faction_id,
-    story,
-    stats
-  } = req.body;
-
-  // validating  stats
-  checkStatsValidity(stats, res);
-  // ...
-  // Heavy lifting
-  const query = {
-    text:
-      "INSERT INTO characters( name, weapon_type_id, img_url, quote, faction_id, story, stats ) VALUES($1, $2, $3, $4, $5, $6, $7)",
-    values: [name, weapon_type_id, img_url, quote, faction_id, story, stats]
-  };
-
-  // promise
-  pool
-    .query(query)
-    .then(response => {
-      console.log(response);
-      res.status(201).json({
-        message: "Character created sucessfully"
-      });
-    })
-    .catch(err => {
-      console.error(err.stack);
-      res.status(500).json({
-        error: err
-      });
-    });
-};
-
-exports.characters_update_character = (req, res, next) => {
-  let query = "update characters set ";
-  const values = [];
-  const prop = req.body;
-
-  for (let i = 0; i < prop.length; i++) {
-    if (prop[i].propName === "stats") {
-      const stats = prop[i].value;
-      // validating  stats
-      checkStatsValidity(stats, res);
-    }
-
-    query += prop[i].propName + "=$" + (i + 1);
-
-    // if is not the last input put a comma after
-    query += i < prop.length - 1 ? "," : " ";
-
-    values[i] = prop[i].value;
-  }
-
-  query += "where id=$" + (prop.length + 1);
-  values[prop.length] = req.params.characterId;
-
-  pool
-    .query(query, values)
-    .then(response => {
-      res.status(200).json({ message: "Character updated sucessfully" });
-    })
-    .catch(err => {
-      console.log(err);
-      res.status(500).json({ error: err });
-    });
-};
-
-exports.characters_delete_character = (req, res, next) => {
-  const id = parseInt(req.params.characterId);
-
-  pool
-    .query("delete from characters where id = $1", [id])
-    .then(result => {
-      res.status(200).json({
-        status: "Character was deleted.",
-        request: {
-          type: "POST",
-          message: "You can also create a new character.",
-          url: "https://node-rest-surge-cards.herokuapp.com/characters/",
-          help: {
-            name: "STRING (req)",
-            weapon_type_id:
-              "id (req)(ref: 'https://node-rest-surge-cards.herokuapp.com/info/characters/weaponType')",
-            img_url: "STRING (req)",
-            quote: "STRING (req)",
-            faction_id:
-              "id (req) (ref: 'https://node-rest-surge-cards.herokuapp.com/info/characters/factions')",
-            story: 'STRING (def: "")',
-            stats: {
-              swp: "NUMBER (req)",
-              shp: "NUMBER (req)",
-              effectName: 'STRING (def: "")',
-              effectDesc: 'STRING (def: "")'
-            }
-          }
-        }
-      });
-    })
-    .catch(function(err) {
-      return res.status(500).json(err);
-    });
 };
